+++ src/main/java/com/hexvane/pathfinder/PathfinderMarkerProvider.java
@@ -2,10 +2,9 @@ package com.hexvane.pathfinder;
 
 import com.hypixel.hytale.protocol.packets.worldmap.MapMarker;
 import com.hypixel.hytale.server.core.entity.entities.Player;
-import com.hypixel.hytale.server.core.entity.entities.player.data.PlayerWorldData;
 import com.hypixel.hytale.server.core.universe.world.World;
 import com.hypixel.hytale.server.core.universe.world.worldmap.WorldMapManager;
-import com.hypixel.hytale.server.core.universe.world.worldmap.markers.MapMarkerTracker;
+import com.hypixel.hytale.server.core.universe.world.worldmap.markers.MarkersCollector;
 import javax.annotation.Nonnull;
 
 public class PathfinderMarkerProvider implements WorldMapManager.MarkerProvider {
@@ -15,23 +14,13 @@ public class PathfinderMarkerProvider implements WorldMapManager.MarkerProvider 
     }
 
     @Override
-    public void update(
-            @Nonnull World world,
-            @Nonnull MapMarkerTracker tracker,
-            int chunkViewRadius,
-            int playerChunkX,
-            int playerChunkZ
-    ) {
-        Player player = tracker.getPlayer();
-        PlayerWorldData perWorldData = player.getPlayerConfigData().getPerWorldData(world.getName());
-        MapMarker[] worldMapMarkers = perWorldData.getWorldMapMarkers();
-        if (worldMapMarkers != null) {
-            for (MapMarker marker : worldMapMarkers) {
-                // Only handle pathfinder markers
-                if (marker.id != null && marker.id.startsWith("pathfinder_")) {
-                    // Use -1 for unlimited visibility (always show regardless of distance)
-                    tracker.trySendMarker(-1, playerChunkX, playerChunkZ, marker);
-                }
+    public void update(@Nonnull World world, @Nonnull Player player, @Nonnull MarkersCollector collector) {
+        // Use Player as key - same instance is used when adding markers from command/GUI
+        MapMarker[] markers = PathfinderMarkerStorage.getMarkers(world.getName(), player);
+        for (MapMarker marker : markers) {
+            if (marker.id != null && marker.id.startsWith("pathfinder_")) {
+                // Use addIgnoreViewDistance so far-away biome markers still show on the map/compass
+                collector.addIgnoreViewDistance(marker);
             }
         }
     }